Added localization plugin and logic
We do not use the localization provided by feeling responsive (allows only the publishing of the website in a specified language).
Instead, we use a plugin which generates the websites in a predefined list of languages.
We use the modified version of the plugin so that the default language (en) is in its iwn subfolder (/en/...).
See https://www.klaasnotfound.com/2017/02/16/proper-multilingual-site-with-github-pages-and-jekyll/