Move build to docs folder for production deployment
We use a modified version of the localization plugin, which is not supported by Github pages.
As a result, the only way to use Github Pages hosting is to build the site locally and push the static files to the docs folder. Github Pages will then deploy based on the docs folder content.